moved profile model to b_m app to match template
